requirements.txt
Flask==2.3.3
Flask-Bcrypt==1.0.1
mysql-connector-python==8.1.0
werkzeug==2.3.7
validators==0.20.0
qrcode==8.0
Pillow==10.3.0
requests==2.32.3
python-dotenv==1.0.0
gunicorn==20.1.0
url-parser==3.0.3